:tada: Start: week8 (rename Folder.js to Shared.js)
Start week8 (rename Folder.js to Shared.js)
